+++ src/main/java/fr/estecka/variantscit/format/RegexTransform.java
@@ -10,13 +10,15 @@ import com.mojang.serialization.MapCodec;
 import com.mojang.serialization.codecs.RecordCodecBuilder;
 import fr.estecka.variantscit.VariantsCitMod;
 
-public record RegexTransform(Pattern pattern, String substitution)
+public record RegexTransform(Pattern pattern, String substitution, boolean matchAll, boolean validate)
 implements IStringTransform
 {
 	static public final MapCodec<RegexTransform> MAPCODEC = RecordCodecBuilder.mapCodec(instance ->
 		instance.group(
 			Codec.STRING.comapFlatMap(RegexTransform::ParsePattern, Pattern::toString).fieldOf("regex").forGetter(RegexTransform::pattern),
-			Codec.STRING.optionalFieldOf("substitution", "$0").forGetter(RegexTransform::substitution)
+			Codec.STRING.optionalFieldOf("substitution", "$0").forGetter(RegexTransform::substitution),
+			Codec.BOOL.optionalFieldOf("matchAll", true).forGetter(RegexTransform::matchAll),
+			Codec.BOOL.optionalFieldOf("validate", true).forGetter(RegexTransform::validate)
 		).apply(instance, RegexTransform::new)
 	);
 
@@ -32,7 +34,7 @@ implements IStringTransform
 	@Override
 	public String apply(String input) {
 		Matcher match = this.pattern.matcher(input);
-		if (match.matches())
+		if (matchAll ? match.matches() : match.find())
 		try {
 			return match.replaceAll(this.substitution);
 		}
@@ -40,6 +42,6 @@ implements IStringTransform
 			VariantsCitMod.LOGGER.error("Error in regex substitution: {}\n{}]", this.substitution, ExceptionUtils.getStackTrace(e));
 		}
 
-		return null;
+		return validate ? null : input;
 	}
 }